requirements.txt
# BikeJJ - Dependências Python
# Versão: 3.0.0 - Arduino Mega + GSAP + Reset Robusto

# Comunicação Serial com Arduino
pyserial>=3.5

# Servidor HTTP
# (http.server é built-in no Python 3.x)

# Utilitários (opcional)
# requests>=2.25.1  # Para testes HTTP
# flask>=2.0.0      # Alternativa ao servidor built-in

# Desenvolvimento (opcional)
# pytest>=6.0.0     # Para testes
# black>=21.0.0     # Formatação de código
# flake8>=3.8.0     # Linting

# Nota: O servidor robust_server.py usa apenas bibliotecas padrão Python
# exceto pyserial para comunicação com Arduino
